+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@ Release Notes
 * ServiceBus: Fix an issue with Premium Sku ARM Writer
 * ServiceBus: Fix an issue with Rules depends on ARM Writer
 * Storage Accounts: Support for CORS.
+* Virtual WAN: add builder
 
 ##  1.5.0
 * Container Groups: Support for init containers.